pyproject.toml
[project]
name = "truth-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.14"
dependencies = [
    "beautifulsoup4",
    "camoufox",
    "fastapi[standard]",
    "httpx",
    "lxml",
    "openai",
    "pillow",
    "pydantic-ai[openrouter,openai,bedrock]",
    "python-dotenv",
    "scrapling[all]",
    # "stable-diffusion-cpp-python; sys_platform == 'linux'",
    "pynvml; sys_platform == 'linux'",
]
